Add microbatch strategy (#924)
* Add microbatch strategy

This work is basically in entirety a duplicate of the work done by
MichelleArk in https://github.com/dbt-labs/dbt-snowflake/pull/1179.
I don't really expect this to work first try, but it might. I expect
to need to do some edits, but who knows, maybe I'll get lucky.

* Add changie doc

* Add comment to microbatch macro to explain why we are re-implementing delete+insert

* Add `begin` to microbatch config in test_incremental_microbatch.py

* Cleanup predicates in microbatch materialization

* Fix predicate/incremental predicate logic in microbatch macro

* Remove unnecessary `if` in microbatch macro

The `if` is unnecessary because predicates are guaranteed to exist,
but the `if` was guarding against when there are no predicates.

* Get batch start and end time in the same way

* Remove unnecessary `target` specifications for columns of predicates in microbatch materialization

The `target.` portion of `target.<column_name>` is unnecessary for the predicates in the
microbatch materialization macro because the delete statement already ensures the "targeting`
of `target` in the delete statement via the clause `delete from {{ target }}`. Said another way,
there is no use of the word `using` in the delete clause, thus it is unambiguous what is being
deleted from.

---------

Co-authored-by: Michelle Ark <MichelleArk@users.noreply.github.com>
Co-authored-by: Mike Alfare <13974384+mikealfare@users.noreply.github.com>