Refactor to use new batch context varaibles
During the betas/rcs of dbt-core 1.9.0, the varaibles `__dbt_internal_event_time_start` and `__dbt_internal_event_time_end` were available on the model in the jinja context. However when the final of dbt-core 1.9.0 was released, dbt-core started providing a `batch` object on the model available in the jinja context. This batch object made availble all the prevous information and more, and it is the "correct" way to get the information (i.e. `__dbt_internal_event_time_end` and `__dbt_internal_event_time_start` are deprecated.